LPS-83841 SF - rename controller to router since this name matches better with apio naming convention
