+++ modules/apps/headless-apio/structured-content/structured-content-apio-api/bnd.bnd
@@ -2,8 +2,8 @@ Bundle-Name: Liferay Structured Content Apio API
 Bundle-SymbolicName: com.liferay.structured.content.apio.api
 Bundle-Version: 2.0.0
 Export-Package:\
-	com.liferay.structured.content.apio.architect.controller,\
 	com.liferay.structured.content.apio.architect.form,\
 	com.liferay.structured.content.apio.architect.identifier,\
 	com.liferay.structured.content.apio.architect.model,\
+	com.liferay.structured.content.apio.architect.router,\
 	com.liferay.structured.content.apio.architect.util
